convert the values to be sent and delete unnecessary console.logs
