+++ src/utils/ADempiere/valueUtils.js
@@ -558,3 +558,56 @@ export function formatConversionCurrenty(params) {
   }
   return params
 }
+/**
+ * convert Values To Send
+ * @param {string, number, boolean, date} values
+ */
+
+export function convertValuesToSend(values) {
+  const valuesToSend = {}
+
+  values.forEach(element => {
+    const { value, columnName } = element
+
+    if (isEmptyValue(value)) {
+      return
+    }
+
+    switch (columnName) {
+      case 'DocumentNo':
+        valuesToSend['documentNo'] = value
+        break
+      case 'C_BPartner_ID_UUID':
+        valuesToSend['businessPartnerUuid'] = value
+        break
+      case 'GrandTotal':
+        valuesToSend['grandTotal'] = value
+        break
+      case 'OpenAmt':
+        valuesToSend['openAmount'] = value
+        break
+      case 'IsPaid':
+        valuesToSend['isPaid'] = value
+        break
+      case 'Processed':
+        valuesToSend['isProcessed'] = value
+        break
+      case 'IsAisleSeller':
+        valuesToSend['isAisleSeller'] = value
+        break
+      case 'IsInvoiced':
+        valuesToSend['isInvoiced'] = value
+        break
+      case 'DateOrderedFrom':
+        valuesToSend['dateOrderedFrom'] = value
+        break
+      case 'DateOrderedTo':
+        valuesToSend['dateOrderedTo'] = value
+        break
+      case 'SalesRep_ID_UUID':
+        valuesToSend['salesRepresentativeUuid'] = value
+        break
+    }
+  })
+  return valuesToSend
+}
